src/main/java/de/codingchallenge/controllers/CategoryController.java
package de.codingchallenge.controllers;

import de.codingchallenge.model.Category;
import de.codingchallenge.repository.CategoryRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

@RestController
@RequestMapping("/api")
public class CategoryController {

	private final CategoryRepository categoryRepository;

	@Autowired
	public CategoryController(CategoryRepository categoryRepository) {
		this.categoryRepository = categoryRepository;
	}

	@RequestMapping("/category")
	public List<Category> categories() {
		return categoryRepository.findAll();
	}
}
